requirements.txt
Kivy~=2.3.1
pillow~=11.1.0